+++ sporco/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-__version__ = '0.1.8b1'
+__version__ = '0.1.8'
 
 import sporco.cdict
 import sporco.util