+++ CHANGES.rst
@@ -9,6 +9,8 @@ Version 0.1.8   (not yet released)
 - Added parallel processing implementation of convolutional dictionary
   learning algorithm based on the hybrid Mask Decoupling/Consensus
   dictionary update to module parcnsdl
+- Fixed bug in package setup that resulted in example image files being
+  omitted from package distributions